+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/RestAPI/Credentials/credentialsDataGrid/generateCredentialsDialog/GeneratedCredentialsSection.tsx
