+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/eventStreams/EventStreamsHeader.tsx
