+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/eventStreams/ResourcesSection.tsx
