+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/eventStreams/serviceAccountsSection/index.tsx
