+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeData/eventStreams/serviceAccountsSection/serviceAccountsDataGrid/viewCredentials/ViewCredentialsDialogContent.tsx
